% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste.R
\name{teste}
\alias{teste}
\title{Teste}
\usage{
teste(gencode_transcripts_path)
}
\arguments{
\item{gencode_transcripts_path}{The pathway of the gencode transcripts fasta file in quotes.}
}
\value{
A table with transcripts and genes identification named "gencode_header.csv" with the follow columns:
\item transcript_id
\item gene_id
\item transcript_name
\item gene_name
\item transcript_biotype
}
\description{
\code{teste} is used to generate a file with the header information of a fasta gencode file
}
\examples{
teste("example.fasta")
}
